Move default typeclass impl. in `*Class` companion.
