+++ base/src/main/scala/data/ConstInstances.scala
@@ -3,8 +3,11 @@ package data
 
 import scalaz.typeclass._
 
+import FoldableClass._
+import TraversableClass._
+
 trait ConstInstances {
-  implicit def traverse[R]: Traversable[Const[R, ?]] = new TraversableClass[Const[R, ?]] with Foldable.FoldRight[Const[R, ?]] with Traversable.Traverse[Const[R, ?]] {
+  implicit def traverse[R]: Traversable[Const[R, ?]] = new TraversableClass[Const[R, ?]] with FoldRight[Const[R, ?]] with Traverse[Const[R, ?]] {
     def map[A, B](ma: Const[R, A])(f: A => B): Const[R, B] = ma.retag
 
     def traverse[F[_], A, B](ta: Const[R, A])(f: A => F[B])(implicit F: Applicative[F]): F[Const[R, B]] =
