+++ base/src/main/scala/typeclass/Strong.scala
@@ -10,16 +10,6 @@ trait Strong[P[_, _]] {
 }
 
 object Strong extends StrongInstances {
-  trait First[P[_, _]] extends Alt[First[P]]{ self: Strong[P] =>
-    override def first[A, B, C](pab: P[A, B]): P[(A, C), (B, C)]
-    override def second[A, B, C](pab: P[A, B]): P[(C, A), (C, B)] =
-      profunctor.dimap[(A, C), (B, C), (C, A), (C, B)](first(pab))(_.swap)(_.swap)
-  }
-  trait Second[P[_, _]] extends Alt[Second[P]]{ self: Strong[P] =>
-    override def second[A, B, C](pab: P[A, B]): P[(C, A), (C, B)]
-    override def first[A, B, C](pab: P[A, B]): P[(A, C), (B, C)] =
-      profunctor.dimap[(C, A), (C, B), (A, C), (B, C)](second(pab))(_.swap)(_.swap)
-  }
-  trait Alt[D <: Alt[D]] { self: D => }
+  def apply[F[_, _]](implicit F: Strong[F]): Strong[F] = F
 }
