+++ base/src/main/scala/typeclass/Traversable.scala
@@ -10,16 +10,5 @@ trait Traversable[T[_]] {
 }
 
 object Traversable extends TraversableInstances with TraversableSyntax {
-
-  trait Traverse[T[_]] extends Alt[Traverse[T]] { self : Traversable[T] =>
-    override def traverse[F[_], A, B](ta: T[A])(f: A => F[B])(implicit F: Applicative[F]): F[T[B]]
-    override def sequence[F[_], A](ta: T[F[A]])(implicit F: Applicative[F]): F[T[A]] = traverse(ta)(identity)
-  }
-  trait Sequence[T[_]] extends Alt[Sequence[T]] { self : Traversable[T] =>
-    override def sequence[F[_], A](ta: T[F[A]])(implicit F: Applicative[F]): F[T[A]]
-    override def traverse[F[_], A, B](ta: T[A])(f: A => F[B])(implicit F: Applicative[F]): F[T[B]] = sequence(functor.map(ta)(f))
-  }
-  trait Alt[D <: Alt[D]] { self: D => }
-
   def apply[T[_]](implicit T: Traversable[T]): Traversable[T] = T
 }
