+++ base/src/main/scala/typeclass/TraversableInstances.scala
@@ -2,9 +2,11 @@ package scalaz
 package typeclass
 
 import Prelude._
+import FoldableClass._
+import TraversableClass._
 
 trait TraversableInstances {
-  implicit val list: Traversable[List] = new TraversableClass[List] with Traversable.Traverse[List] with Foldable.FoldRight[List] {
+  implicit val list: Traversable[List] = new TraversableClass[List] with Traverse[List] with FoldRight[List] {
 
     override def traverse[F[_], A, B](ta: List[A])(f: A => F[B])(implicit F: Applicative[F]): F[List[B]] =
       ta.foldLeft[F[List[B]]](List.empty[B].pure[F]) { (flb, a) => flb.ap(f(a).map(b => (xs: List[B]) => b::xs)) }
